Rename `destructor_tree` files to `scope_tree`
To match the new class names.
