+++ src/goto-programs/Makefile
@@ -5,7 +5,6 @@ SRC = allocate_objects.cpp \
       class_identifier.cpp \
       compute_called_functions.cpp \
       destructor.cpp \
-      destructor_tree.cpp \
       elf_reader.cpp \
       ensure_one_backedge_per_target.cpp \
       format_strings.cpp \
@@ -60,6 +59,7 @@ SRC = allocate_objects.cpp \
       rewrite_union.cpp \
       resolve_inherited_component.cpp \
       safety_checker.cpp \
+      scope_tree.cpp \
       set_properties.cpp \
       show_goto_functions.cpp \
       show_goto_functions_json.cpp \
