+++ src/goto-programs/scope_tree.h
